REFACTOR: migrate Icon component test files to typescript
